refactor(logic): split army_migration relocation into part-of file (#3290)
Decompose `world/army_migration.dart` (516 physical lines) into the army
construction/reconciliation concern (main library file, 298 lines) and a
behaviour-preserving `part of` file `army_migration_relocation.dart`
(227 lines) holding `updateArmyStation` and the per-regiment relocation
helpers. Same library scope, so imports, constants, and shared private
helpers are unchanged.

Remove the file from `tool/logic_source_file_size_baseline.json` so the
Phase-0 source-file-size gate (=500 physical lines) now enforces it.

Refs #3290
