+++ tool/logic_source_file_size_baseline.json
@@ -6,7 +6,6 @@
   "packages/colonizethis_logic/lib/src/orders/order_suggestion_move_army.dart",
   "packages/colonizethis_logic/lib/src/setup/game_setup_create.dart",
   "packages/colonizethis_logic/lib/src/setup/locked_province_assigner.dart",
-  "packages/colonizethis_logic/lib/src/world/army_migration.dart",
   "packages/colonizethis_logic/lib/src/world/capital_and_gp_fall.dart",
   "packages/colonizethis_logic/lib/src/world/fog_resolution.dart"
 ]
